Updated types to speedy. 32bits OS's returned. Thanks God and Mary and Cia!! Enjoy!! :-D
