+++ adare_net/config/adare_net_config.gpr
@@ -1,6 +1,6 @@
 --  Configuration for adare_net generated by Alire
 abstract project Adare_Net_Config is
-   Crate_Version := "0.1.0-dev";
+   Crate_Version := "0.1.1-dev";
    Crate_Name := "adare_net";
 
    Alire_Host_OS := "linux";
@@ -15,6 +15,7 @@ abstract project Adare_Net_Config is
            ,"-gnatn" -- Enable inlining
            ,"-ffunction-sections" -- Separate ELF section for each function
            ,"-fdata-sections" -- Separate ELF section for each variable
+           ,"-gnaty2abCefhiklnprstmM135"
            ,"-gnat12" -- Ada 2012 Mode
           );
