+++ adare_net/src/AUTHORS
@@ -4,4 +4,4 @@
 ===============================================================
 
 CopyLeft   2020-present  Daniel Norte de Moraes  <danielcheagle@tutanota.com>
-CopyLeft   2020-present  Renan Oliveira  <renanoliveira1988@gmail.com>
+CopyLeft   2020-2023  Renan Oliveira  <renanoliveira1988@gmail.com>
