+++ adare_net_examples/config/adare_net_examples_config.gpr
@@ -1,7 +1,7 @@
 --  Configuration for adare_net_examples generated by Alire
 with "adare_net.gpr";
 abstract project Adare_Net_Examples_Config is
-   Crate_Version := "0.1.0-dev";
+   Crate_Version := "0.1.1-dev";
    Crate_Name := "adare_net_examples";
 
    Alire_Host_OS := "linux";
